Separated LedTask
