+++ BlinkUPanas/main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "gpio_example_main.cpp" "taskClass.cpp"
+idf_component_register(SRCS "LedTask.cpp" "gpio_example_main.cpp" "taskClass.cpp"
                     INCLUDE_DIRS ".")
